kdbus: drop superfluous 'struct kdbus_name_list' (ABI break)
We always return the size of any generated slice in the command structure.
There's no need to include it in the payload itself. Drop kdbus_name_list
and return the offset of the array directly.

While at it, drop 'name' from all those structures and commands. This
ioctl does not only dump names, but also unique IDs and activators. Lets
just call it 'CMD_LIST', which is more catchy, anyway.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>