handle: fix EBADFD vs ENOTTY
Right now, our error-returns for invalid ioctls is not consistent.
Dependent on the stage we catch it, we might return EBADFD or ENOTTY. Fix
this by catching invalid ioctls early and always return ENOTTY. If an
ioctl is called on the wrong FD type, we now consistently return EBADFD.

Downside is, we need to hard-code the ioctl-list twice: Once to catch
ENOTTY, and once to actually dispatch the handler. However, this should be
just fine.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>