kdbus: make sure KDBUS_FLAG_NEGOTIATE doesn't change state
So far we handle KDBUS_FLAG_NEGOTIATE but still continued with the ioctl.
Make sure we don't actually execute the ioctl if the flag is given,
otherwise it's mostly useless.

This patch changes kdbus_args_parse() to return >0 if NEGOTIATE is set.
All the ioctl handlers now return for 'ret != 0' and we turn >0 into 0 in
kdbus_handle_ioctl(). This way, NEGOTIATE is handled the same way by all
ioctls, but we leave them the freedom to actually do more than the common
set in kdbus_args_parse(). This will be used by SEND in a follow-up to
make sure we actually perform negotiation on the message, too.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>