reply: drop misleading comment
Callers really must NOT check atomic counters before calling into
kdbus_reply_find(). It is highly racy as atomics do not provide any
memory-barriers. Therefore, you really cannot rely on it being in-sync
with the actual list.

Furthermore, we never optimize error-paths. The situation where we want to
lookup a reply, but there is none, is always an error-path, except for
interrupted SYNC-replies. However, in those cases we shouldn't match on
the destination, anyway. So no reason to optimize this.

Last but not least, if the list is empty, the kdbus_reply_find() call is
cheap, anyway. The connection-lock is what's heavy, but it has to be taken
during SEND at all times, so we should rather optimise the call-side.

Drop the hint in reply.c. The call-side will be fixed later on.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>