pool: fix handling of empty slices
Right now, the slice allocator silently accepts size==0 and returns a new
slice for this. This will cause us troubles as its offset might be the
same as for other slices (due to size==0).

Fix this by preventing slices with size==0 and instead add a new function
that can return fake-slices with size==0. This function always returns
offset==pool->size and size==0. Furthermore, the slice destructor silently
ignores offset==pool->size.

This way, we have a similar logic to malloc(0), which always returns a
valid, but not neccessarily unique pointer.

Reported-by: Lukasz Skalski <l.skalski@samsung.com>
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>