reply: allow replying=NULL in lookups
During interrupted SYNC calls, there is no guarantee that the target name
didn't move. However, if the reply is already queued, this shouldn't
affect the behavior of the SYNC call. Therefore, always lookup pending
reply objects by the offset only. If it matches, use it, regardless which
destination connection is comes from.

We never give any guarantees on cookie conflicts. It's the reposnibility
of the API user to use unique cookies at all times. Otherwise, SYNC calls
will not be guaranteed to pick up the correct reply during interruptions.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>