metadata, queue: rework allocation logic (ABI break)
Currently, we rely on the pool to have sufficient space available at
RECV time to accommodate a slice that carries the actual message. This
was done because with metadata that is translated to the receiver's
namespaces, we cannot know the exact size of the metadata, and hence
he decided to postpone that action to the very last moment.

Turns out this has a number of sever downsides:

 a) The pool could be exhausted at the time of receiving, and our
    current logic to preserve some space for this is sub-optimal

 b) The more slices in the pool, the more fragmented it becomes,
    which is bad for future allocations

 c) We have to dance around and keep "msg_extra" details in an
    extra buffer for kernel notifications

 d) kmalloc() is considered expensive

Hence, chose a cheaper way out and calculate the space needed for
metadata conservatively. It doesn't matter if we don't use up all
the bytes allocated earlier. A size mismatch will only occur when
messages are sent accross namespace boundaries, which is not a case
considered common. Also, for this to happen at all, we need to
do proper translations of EXE items, which we cannot do out-of-tree
anyway due to a missing __d_path() export.

As the metadata is the last segment in the chain of message items,
we can leave a 'hole' in that area. The actual message payload is
stuffed at an offset after this and referenced directly from the
payload items.

Consequently, this means the payload OFF items are now again relative
to the slice, not to the pool, hence the ABI break.

Signed-off-by: Daniel Mack <daniel@zonque.org>