match: don't lock mdb on destruction
We're about to free 'mdb', nobody else can own a reference to it but us.
There's no reason to lock the object. It's usually never a good idea to
lock objects we're about to destroy, anyway.

This fixes a theoretical dead-lock when we release a queue-entry while
holding a connection-lock (conn->lock, queue_entry_release(),
reply_unref(), conn_unref(), conn_free(), match_db_destroy(), mdb->rwlock)
even though the natural lock order is mdb->rwlock _before_ conn->lock.

As this is only theoretical (nobody can have mdb->rwlock held at this
moment, so we cannot dead-lock), this is not really a serious issue.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>