domain: use 'ida' for user-IDs
Our 'user_idr' manager is used to keep a small ID range for all currently
used users. This is needed to allow efficient accounting for user<->user
interactions as we have to allocate an array for all accounted users.

However, we never use this as lookup. We only need the ID management.
Therefore, switching to 'ida' reduces memory consumption considerably as
we no longer need 64bit per ID, but only 1bit.

While at id, rename user->idr to user->id, as 'idr' is really an
implementation detail that should be left to domain.c.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>