domain: use 'idr' for user-management
Static hashmaps scale really badly if we don't know the number of entries
that will be stored (if too small, it'll be slow, if too big, we waste
memory). Furthermore, hash collisions can be predicted by the caller.

For UID mappings, this might not be relevant, as we don't expect a single
user to have control over several IDs at the same time. However, the
memory argument is still valid. Therefore, use the kernel object of choice
for ID lookups: idr

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>