domain/bus: merge ID allocations
Instead of a separate ID name-space for each object-type, we now merge all
IDs. This has two advantages:
 - IDs are unique and will no longer clash in debug dumps
 - we can get rid of a bunch of counters

Disadvantage is, we now share a single cache-line across the whole domain.
However, we already do that for message-IDs, so if we merge the IDs, the
overhead will probably stay the same (as messages are the objects we
optimize for, anyway).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>