connection: rework quota accounting
Our current quota accounting is flawed. It can be easily exploited, but
also limits the wrong resources. By always sending the max-allowed message
size, we can currently send 32 messages to a peer and occupy all pool but
2^-32 bytes. This is effectively the whole pool.
On the other hand, if we send small messages, we can never send more than
64 to a peer without the peer being scheduled.

A recent bump changed this limit to 128, effectively disabling the quota
check.

This patch reworks the whole quota accounting. Instead of limiting the
number of messages, we now limit the total pool-memory per use and the
total queued FDs. The algorithm is described in kdbus_conn_quota(). It
should provide a fair accounting, while still maximizing resource usage.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>