connection: prevent DROP+USE_PRIORITY
DROP in combination with USE_PRIORITY is not reliable. There's no way the
user knows that no message will be received between PEEK and DROP. Prevent
that combination for now. If it turns out that there will be a use-case
that makes sense, we can allow it later on again.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>