connection: return EAGAIN consistently
If you call RECV with RECV_USE_PRIORITY, we currently return ENOMSG
instead of EAGAIN if no suitable message is found. This is highly
confusing and requires generic RECV helpers to check for one more error
code. The caller already knows whether it called USE_PRIORITY. It now
cannot know whether there's a message with less priority left, but that
information can be easily retrieved by calling RECV again but by passing a
higher priority right away, depending on the use-case.

Imho, making EAGAIN consistent is preferable due to consistency, over this
small amount of information carried by ENOMSG. If in doubt, we can always
add a RETURN flag, which is preferably over custom error-codes, anyway (as
it is reliable).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>