kdbus: replace EOVERFLOW with KDBUS_RECV_RETURN_DROPPED_MSGS
We must not fail returning messages on EOVERFLOW. Otherwise, a message
flush of an unprivileged peer might cause us to repeatedly return
EOVERFLOW without ever reading any message. We highly discourage people
from installing MATCH filters for unprivileged peers, but we need to be
backwards compatible to dbus1 here and support that.

Therefore, drop the idea of EOVERFLOW and instead set the
KDBUS_RECV_RETURN_DROPPED_MSGS flag in case 'dropped_msgs' was non-zero.
This is now done if either a message was returned (ret == 0) or if no
messages is available (ret == -EAGAIN). In all other error cases, we
ignore 'dropped_msgs' to simplify writing generic recv-helpers.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>