names: re-order name_entry members
Keep the same order as in all other entry-structures we have. First the
entry properties, then the links. Also avoid zeroeing memory on allocation
as we initialize all objects explicitly, anyway. memset() is currently one
of our most significant performance bottlenecks according to perf. Avoid
it in fast-paths and pseudo-fast-paths. name_acquire() isn't a fast-path
in particular, but we hold a write-lock on the registry, so it does cause
signifcant stalls on fast-paths.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>