connection: cleanup memory quota overflow check
The memory variable is size_t type, the quota->memory is uint32_t, we
need the quota->memory to be small and efficient as possible. On 64bits
the test works, size_t is 8bytes, but on 32bits CPUs the test will have
no effects since U32_MAX is max unsigned int. Just convert to the plain
old overflow checks that work on unsigned integers.

This is not a real issue just code cleanup, we need that guard there in
case the code changes, but currently you can't trigger it. Passing a big
memory variable that overflows is not possible since you will fail the
pool allocation before reaching this code.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>