connection: move_messages() cannot fail
The only condition returning <0 in move_messages() is a WARN_ON(). And
this is even if the messages are already moved to the correct connection.
Therefore, make move_messages() return 'void' and fix up all callers.

Note that we cannot make move_messages() fail, as it would require us to
fail half-way through the move. This is not properly recoverable, so we
simply increase the 'dropped_msgs' counter in those cases. This was always
the case, this patch just makes it explicit.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>