connection: move wait_reply() into cmd_send()
Right now, kdbus_conn_call() performs a synchronous send *and* waits for
the reply. This is cumbersome, as we don't want to lock the whole
name-registry the entire time. A followup will rework the name-registry
locking, so this is a preparation for this:

Instead of waiting for a reply inside of kdbus_conn_call(), we now return
the reply objects to the caller. This simplifies the error-path and leaves
waiting for a reply to kdbus_cmd_send().

If we don't do this, name-registry locking becomes a nightmare in
kdbus_conn_call(), as we conditionally need to unlock the registry when
waiting, but only if we didn't resume the synchronous call.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>