names: replace kdbus_name_lock() hack with explicit locking
About a year ago, we introduced kdbus_name_lock() and kdbus_name_unlock()
to lookup a name and pin it until we release it again. This has always
been a hack as it returns from a function with a read-lock held.
Furthermore, the behavior was never really obvious. You couldn't lock
multiple names and other parts silently relied on it to lock the whole
registry. Otherwise, we wouldn't have a consistent view of source and
destination names during the whole time of a single SEND transaction.

This patch drops those lock-helpers again and implements explicit locking
on the caller's side. We always wanted to avoid it, as it requires us to
lock registry->rwlock outside of names.c. However, by using the helpers,
we effectively do the same. Therefore, explicit locking makes maintenance
easier and allows reviewers to immediately see what is going on.

Furthermore, we can now fully rely on the name-registry to be locked.
Therefore, we can drop other locks in follow-up patches (like conn->lock
to protect conn->names_list, which always has been bogus).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>