bus: lock registry during broadcasts/eavesdropping
If we don't lock the name-registry during broadcasts or eavesdropping,
each listener might get a different view of the message. This is pretty
bad as we might get monitors with different messages than the system will
behave.

Fix this by always locking the name-registry for a whole message
transaction. This is similar to what we do for unicasts already.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>