names: warn if registry is non-empty on destruction
There cannot be any names during registry destruction. We disconnect all
connections *before* tearing down the registry. Therefore, it must be
empty. Drop the cleanup-code and replace it with a WARN_ON. In case
something is left, we will leak the memory, but not try to clean it up (as
something is seriously wrong, anyway).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>