doc: pool - move high-level before low-level stuff
Also fix some minor typos.

TODO: I don't understand the explanation of how POOL_SIZE is computed. What
does "it must match the value that was previously returned through the
KDBUS_ITEM_BLOOM_PARAMETER item field when the KDBUS_CMD_HELLO ioctl
returned" mean? Also the reference to KDBUS_ITEM_BLOOM_PARAMETER in the
'Description' needs to be more specific, it is not very illuminating as
it stands. I guess it would all be clear if the fields of
'struct kdbus_bloom_parameter' are better documented and a reference added.

Signed-off-by: Tom Gundersen <teg@jklm.no>