connection, queue: split quota decrease handler
For synchronous replies to messages, the queue entry is never actually
placed into the queue of the connection that sent the request in the
first place (and which is now waiting for the reply).

This also means that kdbus_queue_entry_remove() is not called for it,
and so the quota it occupies is never decreased in the connection.

However, as the entry is not actually queued, we can't just call
kdbus_queue_entry_remove() for synchronous replies, as the list
and r/b entry pointers are invalid.

To fix this, split off code from kdbus_queue_entry_remove() into
a new function called kdbus_queue_entry_dec_quota(), and call it
from both kdbus_queue_entry_remove() and kdbus_conn_wait_reply().

This fixes a regression detected with sd-bus, where synchronous
message transfers would reliably lead to -ENOBUFS errors after
some time.

Signed-off-by: Daniel Mack <daniel@zonque.org>