pool: when moving slices do not give out more than the half of remaining space
When moving a slice from a pool to another one, make sure to not give
out more than the half of the remaining space of the target pool.

Usually we need this space to make sure that the pool owner always has some
space left to perform its own operations, and the other half is used for
incomming messages. Follow this and add the same check here too, this way we
prevent accumulated messages that are moved from another connection to
consume the target pool.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>