queue: run quota-check in queue_entry_alloc()
We really must run quota-checks *before* allocating the slice. Therefore,
move the quota-check into queue.c and run it in queue_entry_alloc().

This commit also introduces a new semantic for entry->user. It is
initialized to ERR_PTR() on new slices. Therefore, if IS_ERR(entry->user)
is true, the slice is un-accounted. Whenever we account for the message,
we set it to the kdbus_user object or NULL. This way, we always know
whether a slice is currently accounted for or not.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>