pool: stop BUG()ing
If a function can fail, we should never BUG()! BUG() causes a panic and is
a major PITA to debug. The panic-screen is pretty unreliable on my machine
at least (and it runs i915..), so I usually just get a dead machine
without any error message on BUG().

However, if we can return a failure, we can just safely run WARN_ON() and
return an error code without doing something.

Yes, a WARN_ON() may cause corruptions as we don't stop the machine. But
if there's a bug, it will have had side-effects already, so it's not
really a major concern, anymore.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>