queue: always remove queue-entry on destruction
Instead of requiring the caller to do the queue-entry removal directly, we
now do it implicitly on entry-destruction. All cases that called
queue_entry_remove() were directly followed by queue_entry_free(), so
we're good.

In case of sync-messages, we never queue them, therefore the removal is a
no-op as it properly detects whether a message is queued or not.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>