queue: simplify quota by making it implicit
So the quota is used to account for slices. Therefore, we can bind
quota-accounting to the lifetime of the slice. Instead of letting each
caller maintain the quota, we now do it implicit. That is, on slice
allocation it's added, on slice release it's dropped. On slice move, we
move the quota.

No-one outside of queue.c needs to touch the quota anymore (except for the
kdbus_conn_quota_*() helpers in connection.c).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>