queue: provide generic kdbus_queue_entry_enqueue()
So far, we set entry->reply from connection.c, which is kinda a
layer-violation. Provide a generic helper to enqueue a queue entry.

Furthermore, we want to avoid to directly expose kdbus_queue_entry_add()
(and the already hidden kdbus_queue_entry_remove()). These are internal
helpers which should do nothing more than linking it to, or unlinking it
from, a message queue. Therefore, avoid exposing them.

To avoid confusion in the future, rename them to _link() and _unlink().

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>