pool: distinguish owner-allocated and queue-allocated slices
This is the final fixup for quota accounting. Whenever we reserve space
for the incoming queue, we need to make sure we leave enough room for
KDBUS_CMD_LIST and similar commands, which may be used by the
connection-owner to query the bus state.

Therefore, our comments always stated that the incoming queue occupies at
most 50% of the pool. So far, we haven't implemented that, though. This
patch changes pool.c to distinguish accounted and un-accounted slices. All
slices that are created via ioctls by the pool-owner, are unaccounted.
Those are expected to be released shortly after the ioctl. Therefore, they
get 50% of the pool, but may occupy even more space, if required.

All other slices (which is only the incoming queue) will be limited to 50%
of the pool. They're marked as 'accounted' and properly tracked by the
pool. Our quota helpers can now properly calculate 1/3 of half of the pool
when quota is concerned.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>