kdbus.h: switch back to 64-bit UIDs and GIDs (ABI break)
After some reconsideration, we decided to move the binary protocol
back to 64-bit wide UIDs and GIDs. After all, it should be possible
to redefine [gu]id_t to uint64_t and things should continue to
work. As we want to avoid such data types in kdbus.h, let's move
back to 64-bit values and be safe.

Signed-off-by: Daniel Mack <daniel@zonque.org>